+++ src/pages/[author]/posts/[slug]/index.tsx
@@ -1,7 +1,6 @@
 import { GetStaticPaths, GetStaticProps, InferGetStaticPropsType, NextPage } from "next";
-import axios from "axios";
-import cheerio from "cheerio";
 
+import { GetOgp, DisplayOgp, isNull } from "utils/ogp";
 import { getPosts, getPost } from "utils/api";
 import { DisplayDate, HasPassed } from "components/Date";
 import Layout from "components/Layout";
@@ -9,13 +8,7 @@ import AuthorProfile from "components/AuthorProfile";
 import CustomReactMarkdown from "components/CustomdReactMarkdown";
 import { ArticleInfo } from "types/markdownMeta";
 import { SlugPath } from "types/paths";
-
-type OgpData = {
-  url: string;
-  title: string | null;
-  description: string | null;
-  image: string | null;
-};
+import { OgpData } from "types/ogpData";
 
 export const getStaticPaths: GetStaticPaths<SlugPath> = async () => {
   const posts = getPosts();
@@ -31,123 +24,42 @@ export const getStaticPaths: GetStaticPaths<SlugPath> = async () => {
   };
 };
 
-const GetOgp = async (url: string) => {
-  const info: OgpData = {
-    url,
-    title: null,
-    description: null,
-    image: null,
-  };
-
-  const res = await axios
-    .get(encodeURI(url))
-    .then((response) => response)
-    .catch(() => null);
-  if (res === null) return info;
-
-  const html = res.data;
-  const $ = cheerio.load(html);
-
-  const title = $("title").text();
-  const description = $('meta[property="og:description"]').attr("content");
-  const image = $('meta[property="og:image"]').attr("content");
-
-  info.title = title === undefined ? null : title;
-  info.description = description === undefined ? null : description;
-  info.image = image === undefined ? null : image;
-
-  return info;
-};
-
-const DisplayOgp = (url: string, title: string, description: string, isImage: boolean, image: string) => {
-  if (title === "") return "";
-
-  const domains = url.match(/([a-zA-Z0-9-]*\.)+[a-zA-Z]{2,}/g);
-  let domain: string;
-
-  if (domains === null) domain = "";
-  else if (domains.length === 0) domain = "";
-  else domain = domains[0].toString();
-
-  description.replace(/\r?\n/g, "");
-
-  if (isImage)
-    return `
-  <a href="${url}">
-    <div class="w-full my-2 flex border-solid border-2">
-      <div class="w-4/6 sm:w-5/6 p-2">
-        <p class="w-full sm:text-lg font-medium text-black truncate" style="margin: unset">${title}</p>
-        <p class="w-full font-light text-black truncate" style="margin: unset">${description}</p>
-        <p class="w-full text-sm font-light text-black truncate" style="margin: unset">${domain}</p>
-      </div>
-      <img class="h-full w-2/6 sm:w-1/6 my-auto" style="display: ${isImage}" src="${image}" alt=""/>
-    </div>
-  </a>`;
-
-  return `
-  <a href="${url}">
-    <div class="w-full p-2 my-2 border-solid border-2">
-      <p class="w-full sm:text-lg font-medium text-black truncate" style="margin: unset">${title}</p>
-      <p class="w-full font-light text-black truncate" style="margin: unset">${description}</p>
-      <p class="w-full text-sm font-light text-black truncate" style="margin: unset">${domain}</p>
-    </div>
-  </a>`;
-};
-
 export const getStaticProps: GetStaticProps<ArticleInfo, SlugPath> = async ({ params }) => {
-  const data = getPost(params?.author, params?.slug);
+  const data: ArticleInfo = getPost(params?.author, params?.slug);
 
   const httpURL = /https?:\/\/[\w!?/+\-_~;.,*&@#$%=']+/gim;
 
-  const GetLinks = data.content.match(httpURL);
-  const Links = Array.from(new Set(GetLinks));
-  let OGPs: OgpData[];
-  const contents = data.content.split(/\r?\n/);
+  /* リンクを全取得 -> ogpDataに保存 */
+  const links = Array.from(new Set(data.content.match(httpURL)));
 
-  if (Links.length !== 0) {
-    OGPs = await Promise.all(
-      Links.map(async (item) => {
-        const url = item.match(httpURL);
+  const ogpData: OgpData[] = await Promise.all(links.map(async (item) => GetOgp(item)));
 
-        if (url !== null) {
-          return GetOgp(url[0]);
-        }
+  /* 一行ごとに区切る */
+  data.content = data.content.split(/\r?\n/).reduce((prev, line) => {
+    Array.from(new Set(line.match(httpURL))).forEach((url) => {
+      const ogpIndex = links.indexOf(url);
+      if (ogpIndex === -1) return;
 
-        throw new Error();
-      }),
-    );
-
-    contents.forEach((content, index) => {
-      const GetcontentLinks = content.match(httpURL);
-      const contentLinks = Array.from(new Set(GetcontentLinks));
-      if (contentLinks.length === 0) return;
-
-      contentLinks.forEach((val) => {
-        const ogpIndex = Links.indexOf(val);
-        if (ogpIndex === -1) return;
-
-        const urlParams = new URLSearchParams(val.split("?")[1]);
-        const isOgp = urlParams.has("isogp");
-
-        // isogpがあればOGPを非表示にする
-        if (isOgp) {
-          // urlの変更
-          const replaceVal = val.replace(`isogp=${urlParams.get("isogp")}`, " ");
-          contents[index] = contents[index].replace(val, replaceVal);
-          return;
-        }
-
-        const ogp = OGPs[ogpIndex];
-        const title = ogp.title === null ? "" : ogp.title;
-        const description = ogp.description === null ? "" : ogp.description;
-        const image = ogp.image === null ? "" : ogp.image;
-        const isImage = image !== "";
-        contents[index] = `${contents[index]}\n${DisplayOgp(val, title, description, isImage, image)}\n`;
-      });
+      const urlParams = new URLSearchParams(url.split("?")[1]);
+      if (urlParams.has("isogp")) {
+        /* isogpがある場合, lineからsogpを削除する */
+        line.replace(url, url.replace(`isogp=${urlParams.get("isogp")}`, " "));
+      } else {
+        /* isogpが無い場合, lineにogpを追加 */
+        const ogp = ogpData[ogpIndex];
+        line = `${line}\n${DisplayOgp(
+          url,
+          isNull(ogp.title),
+          isNull(ogp.description),
+          ogp.image !== "",
+          isNull(ogp.image),
+        )}\n`;
+      }
     });
-  }
 
-  data.content = contents.join("\n");
+    /* prev(前回までのline)と今回のlineを結合 */
+    return `${prev}\n${line}`;
+  });
 
   return {
     props: data,
