+++ src/types/ogpData.ts
@@ -1,0 +1,6 @@
+export type OgpData = {
+  url: string;
+  title: string | null;
+  description: string | null;
+  image: string | null;
+};
